SEPERATE SCRIPT CONFIGURATION FROM ACTUAL SCRIPT, DEBUG MODE, EMAIL MODE
The script now reads configuration file named "config.php" that is based
on "config-orig.php". Therefor the configuration can be separated from the
script make it more safe and easily manageable.
There is now a debug mode that can be used for testing. In it the script
outputs errors as well as script results in HTML form. The no-debug mode
does not produce any output (besides possible PHP errors that are the
result of misconfiguration of the script.
There is an e-mail and no-e-mail mode. The first sends newly found caches
by email, and the second does not send email for newly found caches.
There is an option to send mail for situations, where no new caches can be
found. For testing purposes.

closes #1
closes #2