NEW CACHES ARE TRACKED BY THEIR ID, ADDED MORE CACHE PARAMETERS TO OUTPUT
The script now extracts more data from newly set caches. It can collect:
- cache's id,
- owner's name,
- date of the cache,
- name,
- link to the cache.

The output is not optimized yet, but it is a start. Will work on it in the
future.

The way a cache is represented in the database has changed. Now caches are
stored by their unique id. That should prevent them to reappear on the list
once their name changes.

To whoever might be notified of this commit: the change is still in the
developer branch, it will stay here until the output is prettified :) It
is a mess right now, but I have to sleep a bit :)